% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amazon_data.R
\docType{data}
\name{amazon_data}
\alias{amazon_data}
\title{Amazon Customer Dataset}
\format{
## `amazon_data`
A data frame with 500 rows and 2 columns:
\describe{
  \item{staring_rating}{star  rating}
  \item{review}{customer review}
  ...
}
}
\source{
<https://s3.amazonaws.com/amazon-reviews-pds/tsv/index.txt>
}
\usage{
data(amazon_data)
}
\description{
A subset of data from the Amazon customer data
Report ...
}
\keyword{datasets}
