% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.R
\name{cluster_kmeans}
\alias{cluster_kmeans}
\title{Clustering by using Kmeans}
\usage{
cluster_kmeans(data_matrix)
}
\arguments{
\item{data_matrix}{emo_mat file as a parameter}
}
\value{
graph
}
\description{
K-means clustering is an Unsupervised Non-linear algorithm that cluster data based on similarity or similar groups. It chooses the number K clusters
It Select at random K points, and forms centroids and it will Assign each data point to closest centroid that forms K clusters.
from the 10 emotions the k means clustering performs cluster based on positive and negative emotions.
}
\references{
https://www.geeksforgeeks.org/k-means-clustering-in-r-programming/
}
