% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.R
\name{convert_to_sparse_matrix}
\alias{convert_to_sparse_matrix}
\title{converting to sparse matrix}
\usage{
convert_to_sparse_matrix()
}
\value{
sparsematrix
}
\description{
It converts emotion matrix into sparse matrix, sparse matrices can be
useful for computing large-scale applications that dense matrices cannot handle.
}
