% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emo_mat.R
\docType{data}
\name{emo_mat}
\alias{emo_mat}
\title{Emotion matrix}
\format{
## `emo_mat`
A data frame with 500 rows and 2 columns:
\describe{
  \item{staring_rating}{star  rating}
  \item{review}{customer review}
  ...
}
}
\source{
<https://s3.amazonaws.com/amazon-reviews-pds/tsv/index.txt>
}
\usage{
data(emo_mat)
}
\description{
after converting to emotion matirx
}
\keyword{datasets}
