% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.R
\name{get_emotion}
\alias{get_emotion}
\title{get emotions}
\usage{
get_emotion(csv_file)
}
\arguments{
\item{csv_file}{csv file as a parameter}
}
\value{
A data frame
}
\description{
This function works by dividing emotions in an column and they are giving us a total of 10
different emotions, for example anger, eagerness, contempt, fear, joy, sadness, surprise, trust, negativity, and positivity,
by using get nrc function the above 10 emotions will divide based on the positive and negative sentiments
}
\references{
https://medium.com/swlh/exploring-sentiment-analysis-a6b53b026131
}
