% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.R
\name{h_cluster}
\alias{h_cluster}
\title{Hirarchial clustering}
\usage{
h_cluster(emo)
}
\arguments{
\item{emo}{data}
}
\value{
graph
}
\description{
Hierarchical clustering is mostly used for combining nearest clusters into one large cluster here in our package we are combing nearest emotions into one
large cluster to find the nearest clusters h cluster firstly calculate the distance between every pair of observation points and store it in a distance matrix. By plotting our data,
we can see that two emotions like “anger” and “negative” are the nearest emotions they were combined and form a large cluster, after that, it updates its distance calculation
and stores the results in a fresh distance matrix. The process then repeats stages 2 and 3 until all clusters have been combined into a single cluster.
}
\references{
https://www.datacamp.com/tutorial/hierarchical-clustering-R
}
