% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.R
\name{heatmap_visualize}
\alias{heatmap_visualize}
\title{visualizing nmf}
\usage{
heatmap_visualize()
}
\value{
graph
}
\description{
A heatmap uses a warm-to-cool color scheme to graphically represent visitor behavior data as hot and cold regions.
Red is the location with the maximum visitor contact that’s means which emotion is strong (repeated more), and warm colors point to those areas,
while cool colors denote those areas with the least visitor interaction which means which emotion weak (repeated less)
}
