\name{hello}
\alias{hello}
\title{Customer emotions reader package}
\usage{
hello()
}
\description{
Prints 'This project is was implemented by a group of 3 students. The project focuses on: 
Aim 1: Pre-process the data using Term Frequency and Inverse Document Frequency (tfidf) to build an R package for converting face emoticons to words. 
Aim 2: Classifying the words that are a good marker of the sentiments using a correlation
matrix.'.
}
\examples{
hello()
}
