% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.R
\name{matrix_conversion}
\alias{matrix_conversion}
\title{matrix conversion}
\usage{
matrix_conversion(data_file)
}
\arguments{
\item{data_file}{csv file as a parameter}
}
\value{
matrix
}
\description{
Converting each of the 10 emotions data frames into a matrix data frame since
dimension reduction will become easy to perform after the conversion of matrix data frame
}
\references{
https://statisticsglobe.com/convert-data-frame-to-matrix-in-r
}
