% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.R
\name{modeling_nmf}
\alias{modeling_nmf}
\title{modeling nmf}
\usage{
modeling_nmf(rank)
}
\arguments{
\item{rank}{sparsematrix as a parameter}
}
\value{
modeling data
}
\description{
modeling nmf to discover hidden semantic patterns in unstructured collection of documents.
Modeling by using Rank where the value or range of ranks for which NMF is performed.
}
