% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.R
\name{nmf_func}
\alias{nmf_func}
\title{NMF dimension reduction}
\usage{
nmf_func(nmfdim)
}
\arguments{
\item{nmfdim}{equal to rank for NMF}
}
\value{
nmf data
}
\description{
Nonnegative matrix factorization (NMF) is an dimension reduction method.
It has become a widely used tool for the analysis of high dimensional data as
it automatically extracts sparse and meaningful features from a set of nonnegative data vectors.
}
