% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.R
\name{norm_sparse_matrix}
\alias{norm_sparse_matrix}
\title{Normalizing the sparsematrix}
\usage{
norm_sparse_matrix()
}
\value{
normalized data
}
\description{
normalization for sparse matrix which is needed for log convert the data to make it as "normal" as feasible,
improving the validity of the statistical analysis results. In other words, the skewness of our
initial data is reduced or eliminated via the log transformation.
}
\references{
https://satijalab.org/seurat/articles/dim_reduction_vignette.html
}
