% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.R
\name{pca_func}
\alias{pca_func}
\title{PCA dimension reduction}
\usage{
pca_func()
}
\value{
pca data
}
\description{
PCA dimension reduction method which is used to reduce the number of variables in your data by extracting important ones from a large pool.
It reduces the dimension of the data with the aim of retaining as much information as possible.
}
\references{
https://rpubs.com/JanpuHou/278584
}
