% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.R
\name{read_inbuilt_data}
\alias{read_inbuilt_data}
\title{Read inbuilt data set}
\usage{
read_inbuilt_data(data)
}
\arguments{
\item{data}{csv file as a parameter}
}
\value{
A data frame
}
\description{
The function read_inbuilt_data is used for reading our data set.
Before beginning data analysis, it may be best to investigate and clean the data.
Due to improper writing, some words might not be recognized by sentiment analysis. As a result,
in this function, once the data is loaded the function also pre-process the data before converting it to the dataframe.
}
