% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.R
\name{visualizing_pca}
\alias{visualizing_pca}
\title{visualizing pca}
\usage{
visualizing_pca()
}
\value{
graph
}
\description{
biplot visualizes the data into two-dimensional chart that represents the relationship between the rows and columns of a table.
optimally represent any two of the following characteristics: distances between observations. relationships between variables.
}
