% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.R
\name{count_emotions}
\alias{count_emotions}
\title{counting emotions}
\usage{
count_emotions(emotion_file)
}
\arguments{
\item{emotion_file}{a csv file}
}
\value{
A data fram
}
\description{
This  function is used to count the total number of words are present
in each emotion after the words have been divided by 10 emotions
}
