% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EmoReader.R
\docType{package}
\name{EmoReader}
\alias{EmoReader}
\alias{EmoReader-package}
\title{EmoReader Emotion Clustering Library}
\description{
Emo Reader
}
\author{
Diskha Shrestha, Roshan Shrestha, Josephat Oyondi, Akhila
}
